requirements.txt
flask==3.0.*
functions-framework==3.*
google-cloud-bigquery==3.23.0